amazon-nova-samples/multimodal-understanding/sample-apps/01-multimodal-with-helper-libraries/requirements.txt
boto3
botocore>=1.35.74
numpy
matplotlib
streamlit
opencv-python
ffmpeg-python
ipython
sagemaker